Refactor: Modify directory structure
