+++ src/components/Footer/StyledFooter.ts
