+++ src/components/Icon/StyledIcon.ts
