+++ src/components/Main/Main.tsx
@@ -1,7 +1,7 @@
 import React, { MutableRefObject } from 'react';
 import Nav from '~/components/Main/Nav/Nav';
 import Section from '~/components/Main/Section/Section';
-import { StyledMain } from '~/styles/Main/StyledMain';
+import { StyledMain } from './StyledMain';
 import { IQuoteData, IQuotesState, IUserInfo } from '~/types/interface';
 import { CardPositionType, ChevronEventType } from '~/types/type';
