+++ src/components/Main/Nav/NavContentWrapper/NavContentWrapper.tsx
@@ -5,7 +5,12 @@ import Icon from '~/components/Icon/Icon';
 import * as colors from '~/styles/common/colors';
 import { IQuoteData } from '~/types/interface';
 import { BOTTOM_CHEVRON_ICON } from '~/styles/common/iconPath';
-import { ContentController, NavBodyButton, StyledNavContentWrapper } from '~/styles/Main/Nav/NavContentWrapper/StyledNavContentWrapper';
+import { 
+    StyledNavContentWrapper, 
+    ContentController, 
+    NavBodyButton 
+} from './StyledNavContentWrapper';
+
 
 export interface ContentProps {
     title: "History" | "Favorite";
