+++ src/components/Main/Nav/NavHeader/NavHeader.tsx
@@ -6,7 +6,7 @@ import * as sizes from '~/styles/common/sizes';
 import * as colors from '~/styles/common/colors';
 import { StyledButton } from "~/styles/common/StyledButton";
 import { LOGOUT_ICON } from "~/styles/common/iconPath";
-import { StyledNavHeader } from "~/styles/Main/Nav/NavHeader/StyledNavHeader";
+import { StyledNavHeader } from "./StyledNavHeader";
 import { IUserInfo } from "~/types/interface";
 
 interface IProps {
