+++ src/components/Main/Nav/NavHeader/StyledNavHeader.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import * as sizes from '~/styles/common/sizes';
-import { StyledNavBox } from "~/styles/Main/Nav/StyledNav";
+import { StyledNavBox } from "~/components/Main/Nav/StyledNav";
 
 export const StyledNavHeader = styled(StyledNavBox)`
     & > :first-child > :first-child {
