+++ src/components/Main/Section/CardWrapper/StyledCard.ts
