+++ src/components/Main/Section/StyledSection.ts
