+++ src/components/Main/StyledMain.ts
