+++ src/components/UserThumbnail/StyledUserThumbnail.ts
